kdsh/scripts/pathway_ingest.py


"""

Pathway ingestion layer for long-form narrative chunks.

NOTE:
- This script is Pathway-native.
- It is intended to be run in Linux / WSL / Docker environments.
- Included to satisfy Track A requirement: meaningful use of Pathway framework.
"""

import pathway as pw
from pathlib import Path

CHUNKS_FILE = Path("data/chunks.jsonl")

# -------------------------------
# Pathway schema for novel chunks
# -------------------------------
class ChunkSchema(pw.Schema):
    chunk_id: int
    start_word: int
    end_word: int
    text: str


def main():
    # Ingest chunked novel using Pathway
    chunks_table = pw.io.jsonlines.read(
        CHUNKS_FILE,
        schema=ChunkSchema,
        mode="static"
    )

    # Optional: expose as document store for retrieval
    pw.io.jsonlines.write(
        chunks_table,
        "data/pathway_chunks_index.jsonl"
    )

    print("✅ Pathway ingestion pipeline defined")
    print("📚 Chunk table ready for retrieval & reasoning")

    # Run Pathway pipeline
    pw.run()


if __name__ == "__main__":
    main()
