kdsh/requirements.txt
pandas
numpy
tqdm
regex
scikit-learn
requests
transformers
sentencepiece
groq
python-dotenv
